requirements.txt
# Database layers
neo4j==5.21.0
chromadb==0.4.24

# Natural Language Processing
spacy==3.7.5
sentence-transformers==2.7.0

# Data processing
pandas==2.2.2
numpy==1.26.4

# Utilities
python-dateutil==2.9.0.post0
pydantic==2.7.4
pyyaml==6.0.1
click==8.1.7

# Development and testing
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==25.3.0
black==24.4.2
mypy==1.10.1

# Fuzzy matching for entity resolution
fuzzywuzzy==0.18.0
python-Levenshtein==0.25.1
